Simplify the simulation.ts
