+++ src/lib/advisors.ts
@@ -1,0 +1,142 @@
+// Advisor message generation
+
+import {
+  Tile,
+  Stats,
+  ServiceCoverage,
+  AdvisorMessage,
+} from '@/types/game';
+
+// PERF: Generate advisor messages - single pass through grid for all building counts
+export function generateAdvisorMessages(stats: Stats, services: ServiceCoverage, grid: Tile[][]): AdvisorMessage[] {
+  const messages: AdvisorMessage[] = [];
+
+  // PERF: Single pass through grid to collect all building stats
+  let unpoweredBuildings = 0;
+  let unwateredBuildings = 0;
+  let abandonedBuildings = 0;
+  let abandonedResidential = 0;
+  let abandonedCommercial = 0;
+  let abandonedIndustrial = 0;
+  
+  for (const row of grid) {
+    for (const tile of row) {
+      // Only count zoned buildings (not grass)
+      if (tile.zone !== 'none' && tile.building.type !== 'grass') {
+        if (!tile.building.powered) unpoweredBuildings++;
+        if (!tile.building.watered) unwateredBuildings++;
+      }
+      
+      // Count abandoned buildings
+      if (tile.building.abandoned) {
+        abandonedBuildings++;
+        if (tile.zone === 'residential') abandonedResidential++;
+        else if (tile.zone === 'commercial') abandonedCommercial++;
+        else if (tile.zone === 'industrial') abandonedIndustrial++;
+      }
+    }
+  }
+
+  // Power advisor
+  if (unpoweredBuildings > 0) {
+    messages.push({
+      name: 'Power Advisor',
+      icon: 'power',
+      messages: [`${unpoweredBuildings} buildings lack power. Build more power plants!`],
+      priority: unpoweredBuildings > 10 ? 'high' : 'medium',
+    });
+  }
+
+  // Water advisor
+  if (unwateredBuildings > 0) {
+    messages.push({
+      name: 'Water Advisor',
+      icon: 'water',
+      messages: [`${unwateredBuildings} buildings lack water. Build water towers!`],
+      priority: unwateredBuildings > 10 ? 'high' : 'medium',
+    });
+  }
+
+  // Finance advisor
+  const netIncome = stats.income - stats.expenses;
+  if (netIncome < 0) {
+    messages.push({
+      name: 'Finance Advisor',
+      icon: 'cash',
+      messages: [`City is running a deficit of $${Math.abs(netIncome)}/month. Consider raising taxes or cutting services.`],
+      priority: netIncome < -500 ? 'critical' : 'high',
+    });
+  }
+
+  // Safety advisor
+  if (stats.safety < 40) {
+    messages.push({
+      name: 'Safety Advisor',
+      icon: 'shield',
+      messages: ['Crime is on the rise. Build more police stations to protect citizens.'],
+      priority: stats.safety < 20 ? 'critical' : 'high',
+    });
+  }
+
+  // Health advisor
+  if (stats.health < 50) {
+    messages.push({
+      name: 'Health Advisor',
+      icon: 'hospital',
+      messages: ['Health services are lacking. Build hospitals to improve citizen health.'],
+      priority: stats.health < 30 ? 'high' : 'medium',
+    });
+  }
+
+  // Education advisor
+  if (stats.education < 50) {
+    messages.push({
+      name: 'Education Advisor',
+      icon: 'education',
+      messages: ['Education levels are low. Build schools and universities.'],
+      priority: stats.education < 30 ? 'high' : 'medium',
+    });
+  }
+
+  // Environment advisor
+  if (stats.environment < 40) {
+    messages.push({
+      name: 'Environment Advisor',
+      icon: 'environment',
+      messages: ['Pollution is high. Plant trees and build parks to improve air quality.'],
+      priority: stats.environment < 20 ? 'high' : 'medium',
+    });
+  }
+
+  // Jobs advisor
+  const jobRatio = stats.jobs / (stats.population || 1);
+  if (stats.population > 100 && jobRatio < 0.8) {
+    messages.push({
+      name: 'Employment Advisor',
+      icon: 'jobs',
+      messages: [`Unemployment is high. Zone more commercial and industrial areas.`],
+      priority: jobRatio < 0.5 ? 'high' : 'medium',
+    });
+  }
+
+  // Abandonment advisor (data already collected above)
+  if (abandonedBuildings > 0) {
+    const details: string[] = [];
+    if (abandonedResidential > 0) details.push(`${abandonedResidential} residential`);
+    if (abandonedCommercial > 0) details.push(`${abandonedCommercial} commercial`);
+    if (abandonedIndustrial > 0) details.push(`${abandonedIndustrial} industrial`);
+    
+    messages.push({
+      name: 'Urban Planning Advisor',
+      icon: 'planning',
+      messages: [
+        `${abandonedBuildings} abandoned building${abandonedBuildings > 1 ? 's' : ''} in your city (${details.join(', ')}).`,
+        'Oversupply has caused buildings to become vacant.',
+        'Increase demand by growing your city or wait for natural redevelopment.'
+      ],
+      priority: abandonedBuildings > 10 ? 'high' : abandonedBuildings > 5 ? 'medium' : 'low',
+    });
+  }
+
+  return messages;
+}
