+++ src/lib/buildings.ts
@@ -1,0 +1,961 @@
+// Building operations: creation, placement, evolution, and multi-tile handling
+
+import {
+  GameState,
+  Tile,
+  Building,
+  BuildingType,
+  ZoneType,
+  ServiceCoverage,
+  BUILDING_STATS,
+  RESIDENTIAL_BUILDINGS,
+  COMMERCIAL_BUILDINGS,
+  INDUSTRIAL_BUILDINGS,
+} from '@/types/game';
+import { getWaterAdjacency, requiresWaterAdjacency } from './adjacency';
+import { findConnectedBridgeTiles, findAdjacentBridgeTiles } from './bridges';
+
+// Building types that don't require construction (already complete when placed)
+export const NO_CONSTRUCTION_TYPES: BuildingType[] = ['grass', 'empty', 'water', 'road', 'bridge', 'tree'];
+
+// Check if a factory_small at this position would render as a farm
+// This matches the deterministic logic in Game.tsx for farm variant selection
+export function isFarmBuilding(x: number, y: number, buildingType: string): boolean {
+  if (buildingType !== 'factory_small') return false;
+  // Same seed calculation as in Game.tsx rendering
+  const seed = (x * 31 + y * 17) % 100;
+  // ~50% chance to be a farm variant (when seed < 50)
+  return seed < 50;
+}
+
+// Check if a building is a "starter" type that can operate without utilities
+// This includes all factory_small (farms AND small factories), small houses, and small shops
+// All starter buildings represent small-scale, self-sufficient operations that don't need
+// municipal power/water infrastructure to begin operating
+export function isStarterBuilding(x: number, y: number, buildingType: string): boolean {
+  if (buildingType === 'house_small' || buildingType === 'shop_small') return true;
+  // ALL factory_small are starters - they can spawn without utilities
+  // Some will render as farms (~50%), others as small factories
+  // Both represent small-scale operations that can function off-grid
+  if (buildingType === 'factory_small') return true;
+  return false;
+}
+
+export function createBuilding(type: BuildingType): Building {
+  // Buildings that don't require construction start at 100% complete
+  const constructionProgress = NO_CONSTRUCTION_TYPES.includes(type) ? 100 : 0;
+  
+  return {
+    type,
+    level: type === 'grass' || type === 'empty' || type === 'water' ? 0 : 1,
+    population: 0,
+    jobs: 0,
+    powered: false,
+    watered: false,
+    onFire: false,
+    fireProgress: 0,
+    age: 0,
+    constructionProgress,
+    abandoned: false,
+  };
+}
+
+export function createTile(x: number, y: number, buildingType: BuildingType = 'grass'): Tile {
+  return {
+    x,
+    y,
+    zone: 'none',
+    building: createBuilding(buildingType),
+    landValue: 50,
+    pollution: 0,
+    crime: 0,
+    traffic: 0,
+    hasSubway: false,
+  };
+}
+
+// Building sizes for multi-tile buildings (width x height)
+const BUILDING_SIZES: Partial<Record<BuildingType, { width: number; height: number }>> = {
+  power_plant: { width: 2, height: 2 },
+  hospital: { width: 2, height: 2 },
+  school: { width: 2, height: 2 },
+  stadium: { width: 3, height: 3 },
+  museum: { width: 3, height: 3 },
+  university: { width: 3, height: 3 },
+  airport: { width: 4, height: 4 },
+  space_program: { width: 3, height: 3 },
+  park_large: { width: 3, height: 3 },
+  mansion: { width: 2, height: 2 },
+  apartment_low: { width: 2, height: 2 },
+  apartment_high: { width: 2, height: 2 },
+  office_low: { width: 2, height: 2 },
+  office_high: { width: 2, height: 2 },
+  mall: { width: 3, height: 3 },
+  // Industrial buildings - small is 1x1, medium is 2x2, large is 3x3
+  factory_medium: { width: 2, height: 2 },
+  factory_large: { width: 3, height: 3 },
+  warehouse: { width: 2, height: 2 },
+  city_hall: { width: 2, height: 2 },
+  amusement_park: { width: 4, height: 4 },
+  // Parks (new sprite sheet)
+  playground_large: { width: 2, height: 2 },
+  baseball_field_small: { width: 2, height: 2 },
+  football_field: { width: 2, height: 2 },
+  baseball_stadium: { width: 3, height: 3 },
+  mini_golf_course: { width: 2, height: 2 },
+  go_kart_track: { width: 2, height: 2 },
+  amphitheater: { width: 2, height: 2 },
+  greenhouse_garden: { width: 2, height: 2 },
+  marina_docks_small: { width: 2, height: 2 },
+  roller_coaster_small: { width: 2, height: 2 },
+  mountain_lodge: { width: 2, height: 2 },
+  mountain_trailhead: { width: 3, height: 3 },
+  // Transportation
+  rail_station: { width: 2, height: 2 },
+};
+
+// Get the size of a building (how many tiles it spans)
+export function getBuildingSize(buildingType: BuildingType): { width: number; height: number } {
+  return BUILDING_SIZES[buildingType] || { width: 1, height: 1 };
+}
+
+// Get construction speed for a building type (larger buildings take longer)
+// Returns percentage progress per tick
+export function getConstructionSpeed(buildingType: BuildingType): number {
+  const size = getBuildingSize(buildingType);
+  const area = size.width * size.height;
+
+  // Base speed: 24-36% per tick for 1x1 buildings (~3-4 ticks to complete)
+  // Scale down by sqrt of area so larger buildings take proportionally longer:
+  // - 1x1 (1 tile):  24-36% per tick → ~3-4 ticks
+  // - 2x2 (4 tiles): 12-18% per tick → ~6-8 ticks
+  // - 3x3 (9 tiles): 8-12% per tick → ~9-12 ticks
+  // - 4x4 (16 tiles): 6-9% per tick → ~11-16 ticks
+  // Construction takes 30% longer overall (speed reduced by 1/1.3)
+  const baseSpeed = 24 + Math.random() * 12;
+  return (baseSpeed / Math.sqrt(area)) / 1.3;
+}
+
+// Check if a multi-tile building can be placed at the given position
+export function canPlaceMultiTileBuilding(
+  grid: Tile[][],
+  x: number,
+  y: number,
+  width: number,
+  height: number,
+  gridSize: number
+): boolean {
+  // Check bounds
+  if (x + width > gridSize || y + height > gridSize) {
+    return false;
+  }
+
+  // Check all tiles are available (grass or tree only - not water, roads, or existing buildings)
+  // NOTE: 'empty' tiles are placeholders from multi-tile buildings, so we can't build on them
+  // without first bulldozing the entire parent building
+  for (let dy = 0; dy < height; dy++) {
+    for (let dx = 0; dx < width; dx++) {
+      const tile = grid[y + dy]?.[x + dx];
+      if (!tile) return false;
+      // Can only build on grass or trees - roads must be bulldozed first
+      if (tile.building.type !== 'grass' && tile.building.type !== 'tree') {
+        return false;
+      }
+    }
+  }
+
+  return true;
+}
+
+// Check if a multi-tile building can be SPAWNED at the given position
+// This is stricter than canPlaceMultiTileBuilding - it doesn't allow 'empty' tiles
+// because those are placeholders for existing multi-tile buildings
+export function canSpawnMultiTileBuilding(
+  grid: Tile[][],
+  x: number,
+  y: number,
+  width: number,
+  height: number,
+  zone: ZoneType,
+  gridSize: number
+): boolean {
+  if (x + width > gridSize || y + height > gridSize) {
+    return false;
+  }
+  
+  for (let dy = 0; dy < height; dy++) {
+    for (let dx = 0; dx < width; dx++) {
+      const tile = grid[y + dy]?.[x + dx];
+      if (!tile) return false;
+      // Must be in the same zone
+      if (tile.zone !== zone) return false;
+      // Can only spawn on grass or trees
+      // NOT 'empty' - those are placeholders for existing multi-tile buildings
+      if (tile.building.type !== 'grass' && tile.building.type !== 'tree') {
+        return false;
+      }
+    }
+  }
+  
+  return true;
+}
+
+// Footprint helpers for organic growth and merging
+// IMPORTANT: Only allow consolidation of truly empty land (grass, tree).
+// Do NOT include 'empty' tiles - those are placeholders for existing multi-tile buildings!
+// Including 'empty' would allow buildings to overlap with each other during evolution.
+const MERGEABLE_TILE_TYPES = new Set<BuildingType>(['grass', 'tree']);
+
+// Small buildings that can be consolidated into larger ones when demand is high
+const CONSOLIDATABLE_BUILDINGS: Record<ZoneType, Set<BuildingType>> = {
+  residential: new Set(['house_small', 'house_medium']),
+  commercial: new Set(['shop_small', 'shop_medium']),
+  industrial: new Set(['factory_small']),
+  none: new Set(),
+};
+
+function isMergeableZoneTile(
+  tile: Tile, 
+  zone: ZoneType, 
+  excludeTile?: { x: number; y: number },
+  allowBuildingConsolidation?: boolean
+): boolean {
+  // The tile being upgraded is always considered mergeable (it's the source of the evolution)
+  if (excludeTile && tile.x === excludeTile.x && tile.y === excludeTile.y) {
+    return tile.zone === zone && !tile.building.onFire && 
+           tile.building.type !== 'water' && tile.building.type !== 'road';
+  }
+  
+  if (tile.zone !== zone) return false;
+  if (tile.building.onFire) return false;
+  if (tile.building.type === 'water' || tile.building.type === 'road' || tile.building.type === 'bridge') return false;
+  
+  // Always allow merging grass and trees - truly unoccupied tiles
+  if (MERGEABLE_TILE_TYPES.has(tile.building.type)) {
+    return true;
+  }
+  
+  // When demand is high, allow consolidating small buildings into larger ones
+  // This enables developed areas to densify without requiring empty land
+  if (allowBuildingConsolidation && CONSOLIDATABLE_BUILDINGS[zone]?.has(tile.building.type)) {
+    return true;
+  }
+  
+  // 'empty' tiles are placeholders for multi-tile buildings and must NOT be merged
+  return false;
+}
+
+function footprintAvailable(
+  grid: Tile[][],
+  originX: number,
+  originY: number,
+  width: number,
+  height: number,
+  zone: ZoneType,
+  gridSize: number,
+  excludeTile?: { x: number; y: number },
+  allowBuildingConsolidation?: boolean
+): boolean {
+  if (originX < 0 || originY < 0 || originX + width > gridSize || originY + height > gridSize) {
+    return false;
+  }
+
+  for (let dy = 0; dy < height; dy++) {
+    for (let dx = 0; dx < width; dx++) {
+      const tile = grid[originY + dy][originX + dx];
+      if (!isMergeableZoneTile(tile, zone, excludeTile, allowBuildingConsolidation)) {
+        return false;
+      }
+    }
+  }
+  return true;
+}
+
+function scoreFootprint(grid: Tile[][], originX: number, originY: number, width: number, height: number, gridSize: number): number {
+  // Prefer footprints that touch roads for access
+  let roadScore = 0;
+  const offsets = [
+    [-1, 0],
+    [1, 0],
+    [0, -1],
+    [0, 1],
+  ];
+
+  for (let dy = 0; dy < height; dy++) {
+    for (let dx = 0; dx < width; dx++) {
+      const gx = originX + dx;
+      const gy = originY + dy;
+      for (const [ox, oy] of offsets) {
+        const nx = gx + ox;
+        const ny = gy + oy;
+        if (nx >= 0 && ny >= 0 && nx < gridSize && ny < gridSize) {
+          const adjacentType = grid[ny][nx].building.type;
+          if (adjacentType === 'road' || adjacentType === 'bridge') {
+            roadScore++;
+          }
+        }
+      }
+    }
+  }
+
+  // Smaller footprints and more road contacts rank higher
+  return roadScore - width * height * 0.25;
+}
+
+export function findFootprintIncludingTile(
+  grid: Tile[][],
+  x: number,
+  y: number,
+  width: number,
+  height: number,
+  zone: ZoneType,
+  gridSize: number,
+  allowBuildingConsolidation?: boolean
+): { originX: number; originY: number } | null {
+  const candidates: { originX: number; originY: number; score: number }[] = [];
+  // The tile at (x, y) is the one being upgraded, so it should be excluded from the "can't merge existing buildings" check
+  const excludeTile = { x, y };
+
+  for (let oy = y - (height - 1); oy <= y; oy++) {
+    for (let ox = x - (width - 1); ox <= x; ox++) {
+      if (!footprintAvailable(grid, ox, oy, width, height, zone, gridSize, excludeTile, allowBuildingConsolidation)) continue;
+      if (x < ox || x >= ox + width || y < oy || y >= oy + height) continue;
+
+      const score = scoreFootprint(grid, ox, oy, width, height, gridSize);
+      candidates.push({ originX: ox, originY: oy, score });
+    }
+  }
+
+  if (candidates.length === 0) return null;
+  candidates.sort((a, b) => b.score - a.score);
+  return { originX: candidates[0].originX, originY: candidates[0].originY };
+}
+
+export function applyBuildingFootprint(
+  grid: Tile[][],
+  originX: number,
+  originY: number,
+  buildingType: BuildingType,
+  zone: ZoneType,
+  level: number,
+  services?: ServiceCoverage
+): Building {
+  const size = getBuildingSize(buildingType);
+  const stats = BUILDING_STATS[buildingType] || { maxPop: 0, maxJobs: 0, pollution: 0, landValue: 0 };
+
+  for (let dy = 0; dy < size.height; dy++) {
+    for (let dx = 0; dx < size.width; dx++) {
+      const cell = grid[originY + dy][originX + dx];
+      if (dx === 0 && dy === 0) {
+        cell.building = createBuilding(buildingType);
+        cell.building.level = level;
+        cell.building.age = 0;
+        if (services) {
+          cell.building.powered = services.power[originY + dy][originX + dx];
+          cell.building.watered = services.water[originY + dy][originX + dx];
+        }
+      } else {
+        cell.building = createBuilding('empty');
+        cell.building.level = 0;
+      }
+      cell.zone = zone;
+      cell.pollution = dx === 0 && dy === 0 ? stats.pollution : 0;
+    }
+  }
+
+  return grid[originY][originX].building;
+}
+
+// Find the origin tile of a multi-tile building that contains the given tile
+// Returns null if the tile is not part of a multi-tile building
+export function findBuildingOrigin(
+  grid: Tile[][],
+  x: number,
+  y: number,
+  gridSize: number
+): { originX: number; originY: number; buildingType: BuildingType } | null {
+  const tile = grid[y]?.[x];
+  if (!tile) return null;
+  
+  // If this tile has an actual building (not empty), check if it's multi-tile
+  if (tile.building.type !== 'empty' && tile.building.type !== 'grass' && 
+      tile.building.type !== 'water' && tile.building.type !== 'road' && 
+      tile.building.type !== 'bridge' && tile.building.type !== 'rail' && tile.building.type !== 'tree') {
+    const size = getBuildingSize(tile.building.type);
+    if (size.width > 1 || size.height > 1) {
+      return { originX: x, originY: y, buildingType: tile.building.type };
+    }
+    return null; // Single-tile building
+  }
+  
+  // If this is an 'empty' tile, it might be part of a multi-tile building
+  // Search nearby tiles to find the origin
+  if (tile.building.type === 'empty') {
+    // Check up to 4 tiles away (max building size is 4x4)
+    const maxSize = 4;
+    for (let dy = 0; dy < maxSize; dy++) {
+      for (let dx = 0; dx < maxSize; dx++) {
+        const checkX = x - dx;
+        const checkY = y - dy;
+        if (checkX >= 0 && checkY >= 0 && checkX < gridSize && checkY < gridSize) {
+          const checkTile = grid[checkY][checkX];
+          if (checkTile.building.type !== 'empty' && 
+              checkTile.building.type !== 'grass' &&
+              checkTile.building.type !== 'water' &&
+              checkTile.building.type !== 'road' &&
+              checkTile.building.type !== 'bridge' &&
+              checkTile.building.type !== 'rail' &&
+              checkTile.building.type !== 'tree') {
+            const size = getBuildingSize(checkTile.building.type);
+            // Check if this building's footprint includes our original tile
+            if (x >= checkX && x < checkX + size.width &&
+                y >= checkY && y < checkY + size.height) {
+              return { originX: checkX, originY: checkY, buildingType: checkTile.building.type };
+            }
+          }
+        }
+      }
+    }
+  }
+  
+  return null;
+}
+
+// Place a building or zone
+export function placeBuilding(
+  state: GameState,
+  x: number,
+  y: number,
+  buildingType: BuildingType | null,
+  zone: ZoneType | null
+): GameState {
+  const tile = state.grid[y]?.[x];
+  if (!tile) return state;
+
+  // Can't build on water
+  if (tile.building.type === 'water') return state;
+
+  // Can't place roads on existing buildings (only allow on grass, tree, existing roads, or rail - rail+road creates combined tile)
+  // Note: 'empty' tiles are part of multi-tile building footprints, so roads can't be placed there either
+  if (buildingType === 'road') {
+    const allowedTypes: BuildingType[] = ['grass', 'tree', 'road', 'rail'];
+    if (!allowedTypes.includes(tile.building.type)) {
+      return state; // Can't place road on existing building
+    }
+  }
+
+  // Can't place rail on existing buildings (only allow on grass, tree, existing rail, or road - rail+road creates combined tile)
+  if (buildingType === 'rail') {
+    const allowedTypes: BuildingType[] = ['grass', 'tree', 'rail', 'road'];
+    if (!allowedTypes.includes(tile.building.type)) {
+      return state; // Can't place rail on existing building
+    }
+  }
+
+  // Roads, bridges, and rail can be combined, but other buildings require clearing first
+  if (buildingType && buildingType !== 'road' && buildingType !== 'rail' && (tile.building.type === 'road' || tile.building.type === 'bridge')) {
+    return state;
+  }
+  if (buildingType && buildingType !== 'road' && buildingType !== 'rail' && tile.building.type === 'rail') {
+    return state;
+  }
+
+  const newGrid = state.grid.map(row => row.map(t => ({ ...t, building: { ...t.building } })));
+
+  if (zone !== null) {
+    // De-zoning (zone === 'none') can work on any zoned tile/building
+    // Regular zoning can only be applied to grass, tree, or road tiles
+    if (zone === 'none') {
+      // Check if this tile is part of a multi-tile building (handles both origin and 'empty' tiles)
+      const origin = findBuildingOrigin(newGrid, x, y, state.gridSize);
+      
+      if (origin) {
+        // Dezone the entire multi-tile building
+        const size = getBuildingSize(origin.buildingType);
+        for (let dy = 0; dy < size.height; dy++) {
+          for (let dx = 0; dx < size.width; dx++) {
+            const clearX = origin.originX + dx;
+            const clearY = origin.originY + dy;
+            if (clearX < state.gridSize && clearY < state.gridSize) {
+              newGrid[clearY][clearX].building = createBuilding('grass');
+              newGrid[clearY][clearX].zone = 'none';
+            }
+          }
+        }
+      } else {
+        // Single tile - can only dezone tiles that actually have a zone
+        if (tile.zone === 'none') {
+          return state;
+        }
+        // De-zoning resets to grass
+        newGrid[y][x].zone = 'none';
+        newGrid[y][x].building = createBuilding('grass');
+      }
+    } else {
+      // Can't zone over existing buildings (only allow zoning on grass, tree, or road)
+      // NOTE: 'empty' tiles are part of multi-tile buildings, so we can't zone them either
+      const allowedTypesForZoning: BuildingType[] = ['grass', 'tree', 'road'];
+      if (!allowedTypesForZoning.includes(tile.building.type)) {
+        return state; // Can't zone over existing building or part of multi-tile building
+      }
+      // Setting zone
+      newGrid[y][x].zone = zone;
+    }
+  } else if (buildingType) {
+    const size = getBuildingSize(buildingType);
+    
+    // Check water adjacency requirement for waterfront buildings (marina, pier)
+    let shouldFlip = false;
+    if (requiresWaterAdjacency(buildingType)) {
+      const waterCheck = getWaterAdjacency(newGrid, x, y, size.width, size.height, state.gridSize);
+      if (!waterCheck.hasWater) {
+        return state; // Waterfront buildings must be placed next to water
+      }
+      shouldFlip = waterCheck.shouldFlip;
+    }
+    
+    if (size.width > 1 || size.height > 1) {
+      // Multi-tile building - check if we can place it
+      if (!canPlaceMultiTileBuilding(newGrid, x, y, size.width, size.height, state.gridSize)) {
+        return state; // Can't place here
+      }
+      applyBuildingFootprint(newGrid, x, y, buildingType, 'none', 1);
+      // Set flip for waterfront buildings to face the water
+      if (shouldFlip) {
+        newGrid[y][x].building.flipped = true;
+      }
+    } else {
+      // Single tile building - check if tile is available
+      // Can't place on water, existing buildings, or 'empty' tiles (part of multi-tile buildings)
+      // Note: 'road' and 'rail' are included here so they can extend over existing roads/rails,
+      // but non-road/rail buildings are already blocked from roads/rails by the checks above
+      const allowedTypes: BuildingType[] = ['grass', 'tree', 'road', 'rail'];
+      if (!allowedTypes.includes(tile.building.type)) {
+        return state; // Can't place on existing building or part of multi-tile building
+      }
+      
+      // Handle combined rail+road tiles
+      if (buildingType === 'rail' && tile.building.type === 'road') {
+        // Placing rail on road: keep as road with rail overlay
+        newGrid[y][x].hasRailOverlay = true;
+        // Don't change the building type - it stays as road
+      } else if (buildingType === 'road' && tile.building.type === 'rail') {
+        // Placing road on rail: convert to road with rail overlay
+        newGrid[y][x].building = createBuilding('road');
+        newGrid[y][x].hasRailOverlay = true;
+        newGrid[y][x].zone = 'none';
+      } else if (buildingType === 'rail' && tile.hasRailOverlay) {
+        // Already has rail overlay, do nothing
+      } else if (buildingType === 'road' && tile.hasRailOverlay) {
+        // Already has road with rail overlay, do nothing
+      } else {
+        // Normal placement
+        newGrid[y][x].building = createBuilding(buildingType);
+        newGrid[y][x].zone = 'none';
+        // Clear rail overlay if placing non-combined building
+        if (buildingType !== 'road') {
+          newGrid[y][x].hasRailOverlay = false;
+        }
+      }
+      // Set flip for waterfront buildings to face the water
+      if (shouldFlip) {
+        newGrid[y][x].building.flipped = true;
+      }
+    }
+    
+    // NOTE: Bridge creation is handled separately during drag operations across water
+    // We do NOT auto-create bridges here because placing individual road tiles on opposite
+    // sides of water should not automatically create a bridge - only explicit dragging should
+  }
+
+  return { ...state, grid: newGrid };
+}
+
+// Check if a tile has road access by looking for a path through the same zone
+// within a limited distance. This allows large contiguous zones to develop even
+// when only the perimeter touches a road.
+export function hasRoadAccess(
+  grid: Tile[][],
+  x: number,
+  y: number,
+  size: number,
+  maxDistance: number = 8
+): boolean {
+  const startZone = grid[y][x].zone;
+  if (startZone === 'none') {
+    return false;
+  }
+
+  // PERF: Use typed array for visited flags instead of Set<string>
+  // Clear only the area we'll actually use (maxDistance radius)
+  const minClearX = Math.max(0, x - maxDistance);
+  const maxClearX = Math.min(size - 1, x + maxDistance);
+  const minClearY = Math.max(0, y - maxDistance);
+  const maxClearY = Math.min(size - 1, y + maxDistance);
+  for (let cy = minClearY; cy <= maxClearY; cy++) {
+    for (let cx = minClearX; cx <= maxClearX; cx++) {
+      roadAccessVisited[cy * size + cx] = 0;
+    }
+  }
+
+  // BFS using flat queue array [x0, y0, dist0, x1, y1, dist1, ...]
+  let queueHead = 0;
+  let queueTail = 3;
+  roadAccessQueue[0] = x;
+  roadAccessQueue[1] = y;
+  roadAccessQueue[2] = 0;
+  roadAccessVisited[y * size + x] = 1;
+
+  while (queueHead < queueTail) {
+    const cx = roadAccessQueue[queueHead];
+    const cy = roadAccessQueue[queueHead + 1];
+    const dist = roadAccessQueue[queueHead + 2];
+    queueHead += 3;
+    
+    if (dist >= maxDistance) {
+      continue;
+    }
+
+    // Check all 4 directions: [-1,0], [1,0], [0,-1], [0,1]
+    const neighbors = [
+      [cx - 1, cy], [cx + 1, cy], [cx, cy - 1], [cx, cy + 1]
+    ];
+    
+    for (const [nx, ny] of neighbors) {
+      if (nx < 0 || nx >= size || ny < 0 || ny >= size) continue;
+
+      const idx = ny * size + nx;
+      if (roadAccessVisited[idx]) continue;
+      roadAccessVisited[idx] = 1;
+
+      const neighbor = grid[ny][nx];
+
+      if (neighbor.building.type === 'road' || neighbor.building.type === 'bridge') {
+        return true;
+      }
+
+      const isPassableZone = neighbor.zone === startZone && neighbor.building.type !== 'water';
+      if (isPassableZone && queueTail < roadAccessQueue.length - 3) {
+        roadAccessQueue[queueTail] = nx;
+        roadAccessQueue[queueTail + 1] = ny;
+        roadAccessQueue[queueTail + 2] = dist + 1;
+        queueTail += 3;
+      }
+    }
+  }
+
+  return false;
+}
+
+// Evolve buildings based on conditions, reserving footprints as density increases
+export function evolveBuilding(grid: Tile[][], x: number, y: number, services: ServiceCoverage, demand?: { residential: number; commercial: number; industrial: number }): Building {
+  const tile = grid[y][x];
+  const building = tile.building;
+  const zone = tile.zone;
+
+  // Only evolve zoned tiles with real buildings
+  if (zone === 'none' || building.type === 'grass' || building.type === 'water' || building.type === 'road' || building.type === 'bridge') {
+    return building;
+  }
+
+  // Placeholder tiles from multi-tile footprints stay inert but track utilities
+  if (building.type === 'empty') {
+    building.powered = services.power[y][x];
+    building.watered = services.water[y][x];
+    building.population = 0;
+    building.jobs = 0;
+    return building;
+  }
+
+  building.powered = services.power[y][x];
+  building.watered = services.water[y][x];
+
+  const hasPower = building.powered;
+  const hasWater = building.watered;
+  const landValue = tile.landValue;
+  
+  // Starter buildings (farms, house_small, shop_small) don't require power/water
+  const isStarter = isStarterBuilding(x, y, building.type);
+
+  if (!isStarter && (!hasPower || !hasWater)) {
+    return building;
+  }
+
+  // Progress construction if building is not yet complete
+  // Construction requires power and water to progress (except farms)
+  if (building.constructionProgress !== undefined && building.constructionProgress < 100) {
+    // Construction speed scales with building size (larger buildings take longer)
+    const constructionSpeed = getConstructionSpeed(building.type);
+    building.constructionProgress = Math.min(100, building.constructionProgress + constructionSpeed);
+    
+    // While under construction, buildings don't generate population or jobs
+    building.population = 0;
+    building.jobs = 0;
+    
+    // Don't age or evolve until construction is complete
+    return building;
+  }
+
+  // Get zone demand for abandonment/recovery logic
+  const zoneDemandValue = demand ? (
+    zone === 'residential' ? demand.residential :
+    zone === 'commercial' ? demand.commercial :
+    zone === 'industrial' ? demand.industrial : 0
+  ) : 0;
+
+  // === ABANDONMENT MECHANIC ===
+  // Buildings can become abandoned when demand is very negative (oversupply)
+  // Abandoned buildings produce nothing but can recover when demand returns
+  
+  if (building.abandoned) {
+    // Abandoned building - check for recovery
+    // When demand is positive, abandoned buildings have a chance to be cleared
+    // The cleared land (zoned grass) can then be redeveloped
+    if (zoneDemandValue > 10) {
+      // Higher demand = higher chance of clearing abandoned building
+      // At demand 30, ~3% chance per tick; at demand 60, ~8% chance
+      const clearingChance = Math.min(0.12, (zoneDemandValue - 10) / 600);
+      if (Math.random() < clearingChance) {
+        // Clear the abandoned building - revert to zoned grass
+        // This allows natural redevelopment when demand recovers
+        // For multi-tile buildings, clear the entire footprint to avoid orphaned 'empty' tiles
+        const size = getBuildingSize(building.type);
+        if (size.width > 1 || size.height > 1) {
+          // Clear all tiles in the footprint
+          for (let dy = 0; dy < size.height; dy++) {
+            for (let dx = 0; dx < size.width; dx++) {
+              const clearTile = grid[y + dy]?.[x + dx];
+              if (clearTile) {
+                const clearedBuilding = createBuilding('grass');
+                clearedBuilding.powered = services.power[y + dy]?.[x + dx] ?? false;
+                clearedBuilding.watered = services.water[y + dy]?.[x + dx] ?? false;
+                clearTile.building = clearedBuilding;
+              }
+            }
+          }
+        }
+        // Return grass for the origin tile
+        const clearedBuilding = createBuilding('grass');
+        clearedBuilding.powered = building.powered;
+        clearedBuilding.watered = building.watered;
+        return clearedBuilding;
+      }
+    }
+    
+    // Abandoned buildings produce nothing
+    building.population = 0;
+    building.jobs = 0;
+    // Abandoned buildings still age but much slower
+    building.age = (building.age || 0) + 0.1;
+    return building;
+  }
+  
+  // Check if building should become abandoned (oversupply situation)
+  // Only happens when demand is significantly negative and building has been around a while
+  // Abandonment is gradual - even at worst conditions, only ~2-3% of buildings abandon per tick
+  if (zoneDemandValue < -20 && building.age > 30) {
+    // Worse demand = higher chance of abandonment, but capped low for gradual effect
+    // At demand -40, ~0.5% chance per tick; at demand -100, ~2% chance
+    const abandonmentChance = Math.min(0.02, Math.abs(zoneDemandValue + 20) / 4000);
+
+    // Buildings without power/water are slightly more likely to be abandoned (except starter buildings)
+    const utilityPenalty = isStarter ? 0 : ((!hasPower ? 0.005 : 0) + (!hasWater ? 0.005 : 0));
+
+    // Lower-level buildings are slightly more likely to be abandoned
+    const levelPenalty = building.level <= 2 ? 0.003 : 0;
+
+    if (Math.random() < abandonmentChance + utilityPenalty + levelPenalty) {
+      building.abandoned = true;
+      building.population = 0;
+      building.jobs = 0;
+      return building;
+    }
+  }
+
+  building.age = (building.age || 0) + 1;
+
+  // Determine target building based on zone and conditions
+  const buildingList = zone === 'residential' ? RESIDENTIAL_BUILDINGS :
+    zone === 'commercial' ? COMMERCIAL_BUILDINGS :
+    zone === 'industrial' ? INDUSTRIAL_BUILDINGS : [];
+
+  // Calculate level based on land value, services, and demand
+  const serviceCoverage = (
+    services.police[y][x] +
+    services.fire[y][x] +
+    services.health[y][x] +
+    services.education[y][x]
+  ) / 4;
+
+  // Get zone demand to factor into level calculation
+  const zoneDemandForLevel = demand ? (
+    zone === 'residential' ? demand.residential :
+    zone === 'commercial' ? demand.commercial :
+    zone === 'industrial' ? demand.industrial : 0
+  ) : 0;
+  
+  // High demand increases target level, encouraging densification
+  // At demand 60, adds ~0.5 level; at demand 100, adds ~1 level
+  const demandLevelBoost = Math.max(0, (zoneDemandForLevel - 30) / 70) * 0.7;
+
+  const targetLevel = Math.min(5, Math.max(1, Math.floor(
+    (landValue / 24) + (serviceCoverage / 28) + (building.age / 60) + demandLevelBoost
+  )));
+
+  const targetIndex = Math.min(buildingList.length - 1, targetLevel - 1);
+  const targetType = buildingList[targetIndex];
+  let anchorX = x;
+  let anchorY = y;
+
+  // Calculate consolidation probability based on demand
+  // Base probability is low to make consolidation gradual
+  let consolidationChance = 0.08;
+  let allowBuildingConsolidation = false;
+  
+  // Check if this is a small/medium density building that could consolidate
+  const isSmallResidential = zone === 'residential' && 
+    (building.type === 'house_small' || building.type === 'house_medium');
+  const isSmallCommercial = zone === 'commercial' && 
+    (building.type === 'shop_small' || building.type === 'shop_medium');
+  const isSmallIndustrial = zone === 'industrial' && 
+    building.type === 'factory_small';
+  
+  // Get relevant demand for this zone
+  const zoneDemand = demand ? (
+    zone === 'residential' ? demand.residential :
+    zone === 'commercial' ? demand.commercial :
+    zone === 'industrial' ? demand.industrial : 0
+  ) : 0;
+  
+  if (zoneDemand > 30) {
+    if (isSmallResidential || isSmallCommercial || isSmallIndustrial) {
+      // Gradual boost based on demand: at demand 60 adds ~10%, at demand 100 adds ~23%
+      const demandBoost = Math.min(0.25, (zoneDemand - 30) / 300);
+      consolidationChance += demandBoost;
+      
+      // At very high demand (> 70), allow consolidating existing small buildings
+      // but keep the probability increase modest
+      if (zoneDemand > 70) {
+        consolidationChance += 0.05;
+        // Allow consolidating existing small buildings (not just empty land)
+        // This enables developed areas to densify
+        allowBuildingConsolidation = true;
+      }
+    }
+  }
+
+  // Attempt to upgrade footprint/density when the tile is mature enough
+  // Keep consistent age requirement to prevent sudden mass consolidation
+  // Consolidation ALWAYS requires utilities (power and water) - no farm exemption
+  // because consolidation upgrades buildings to larger types that need utilities
+  const ageRequirement = 12;
+  const hasUtilitiesForConsolidation = hasPower && hasWater;
+  if (hasUtilitiesForConsolidation && building.age > ageRequirement && (targetLevel > building.level || targetType !== building.type) && Math.random() < consolidationChance) {
+    const size = getBuildingSize(targetType);
+    const footprint = findFootprintIncludingTile(grid, x, y, size.width, size.height, zone, grid.length, allowBuildingConsolidation);
+
+    if (footprint) {
+      const anchor = applyBuildingFootprint(grid, footprint.originX, footprint.originY, targetType, zone, targetLevel, services);
+      anchor.level = targetLevel;
+      anchorX = footprint.originX;
+      anchorY = footprint.originY;
+    } else if (targetLevel > building.level) {
+      // If we can't merge lots, still allow incremental level gain
+      building.level = Math.min(targetLevel, building.level + 1);
+    }
+  }
+
+  // Always refresh stats on the anchor tile
+  const anchorTile = grid[anchorY][anchorX];
+  const anchorBuilding = anchorTile.building;
+  anchorBuilding.powered = services.power[anchorY][anchorX];
+  anchorBuilding.watered = services.water[anchorY][anchorX];
+  anchorBuilding.level = Math.max(anchorBuilding.level, Math.min(targetLevel, anchorBuilding.level + 1));
+
+  const buildingStats = BUILDING_STATS[anchorBuilding.type];
+  const efficiency = (anchorBuilding.powered ? 0.5 : 0) + (anchorBuilding.watered ? 0.5 : 0);
+
+  anchorBuilding.population = buildingStats?.maxPop > 0
+    ? Math.floor(buildingStats.maxPop * Math.max(1, anchorBuilding.level) * efficiency * 0.8)
+    : 0;
+  anchorBuilding.jobs = buildingStats?.maxJobs > 0
+    ? Math.floor(buildingStats.maxJobs * Math.max(1, anchorBuilding.level) * efficiency * 0.8)
+    : 0;
+
+  return grid[y][x].building;
+}
+
+// Bulldoze a tile (or entire multi-tile building if applicable)
+export function bulldozeTile(state: GameState, x: number, y: number): GameState {
+  const tile = state.grid[y]?.[x];
+  if (!tile) return state;
+  if (tile.building.type === 'water') return state;
+
+  const newGrid = state.grid.map(row => row.map(t => ({ ...t, building: { ...t.building } })));
+  
+  // Special handling for bridges - delete the entire bridge and restore water
+  if (tile.building.type === 'bridge') {
+    const bridgeTiles = findConnectedBridgeTiles(newGrid, state.gridSize, x, y);
+    for (const bt of bridgeTiles) {
+      newGrid[bt.y][bt.x].building = createBuilding('water');
+      newGrid[bt.y][bt.x].zone = 'none';
+      newGrid[bt.y][bt.x].hasRailOverlay = false;
+    }
+    return { ...state, grid: newGrid };
+  }
+  
+  // Special handling for roads - check if adjacent to a bridge start/end
+  if (tile.building.type === 'road') {
+    const adjacentBridgeTiles = findAdjacentBridgeTiles(newGrid, state.gridSize, x, y);
+    if (adjacentBridgeTiles.length > 0) {
+      // Delete the road first
+      newGrid[y][x].building = createBuilding('grass');
+      newGrid[y][x].zone = 'none';
+      newGrid[y][x].hasRailOverlay = false;
+      // Then delete all connected bridge tiles
+      for (const bt of adjacentBridgeTiles) {
+        newGrid[bt.y][bt.x].building = createBuilding('water');
+        newGrid[bt.y][bt.x].zone = 'none';
+        newGrid[bt.y][bt.x].hasRailOverlay = false;
+      }
+      return { ...state, grid: newGrid };
+    }
+  }
+  
+  // Check if this tile is part of a multi-tile building
+  const origin = findBuildingOrigin(newGrid, x, y, state.gridSize);
+  
+  if (origin) {
+    // Bulldoze the entire multi-tile building
+    const size = getBuildingSize(origin.buildingType);
+    for (let dy = 0; dy < size.height; dy++) {
+      for (let dx = 0; dx < size.width; dx++) {
+        const clearX = origin.originX + dx;
+        const clearY = origin.originY + dy;
+        if (clearX < state.gridSize && clearY < state.gridSize) {
+          newGrid[clearY][clearX].building = createBuilding('grass');
+          newGrid[clearY][clearX].zone = 'none';
+          newGrid[clearY][clearX].hasRailOverlay = false; // Clear rail overlay
+          // Don't remove subway when bulldozing surface buildings
+        }
+      }
+    }
+  } else {
+    // Single tile bulldoze
+    newGrid[y][x].building = createBuilding('grass');
+    newGrid[y][x].zone = 'none';
+    newGrid[y][x].hasRailOverlay = false; // Clear rail overlay
+    // Don't remove subway when bulldozing surface buildings
+  }
+
+  return { ...state, grid: newGrid };
+}
+
+// PERF: Pre-allocated arrays for hasRoadAccess BFS to avoid GC pressure
+// Queue stores [x, y, dist] tuples as flat array (3 values per entry)
+const roadAccessQueue = new Int16Array(3 * 256); // Max 256 tiles to check (8*8*4 directions)
+const roadAccessVisited = new Uint8Array(128 * 128); // Max 128x128 grid, reused between calls
+
+// Check if a tile has road access by looking for a path through the same zone
+// within a limited distance. This allows large contiguous zones to develop even
+// when only the perimeter touches a road.
