+++ src/lib/stats.ts
@@ -1,0 +1,302 @@
+// Statistics and budget calculations
+
+import {
+  Tile,
+  Stats,
+  Budget,
+  ServiceCoverage,
+} from '@/types/game';
+
+export function createInitialBudget(): Budget {
+  return {
+    police: { name: 'Police', funding: 100, cost: 0 },
+    fire: { name: 'Fire', funding: 100, cost: 0 },
+    health: { name: 'Health', funding: 100, cost: 0 },
+    education: { name: 'Education', funding: 100, cost: 0 },
+    transportation: { name: 'Transportation', funding: 100, cost: 0 },
+    parks: { name: 'Parks', funding: 100, cost: 0 },
+    power: { name: 'Power', funding: 100, cost: 0 },
+    water: { name: 'Water', funding: 100, cost: 0 },
+  };
+}
+
+export function createInitialStats(): Stats {
+  return {
+    population: 0,
+    jobs: 0,
+    money: 100000,
+    income: 0,
+    expenses: 0,
+    happiness: 50,
+    health: 50,
+    education: 50,
+    safety: 50,
+    environment: 75,
+    demand: {
+      residential: 50,
+      commercial: 30,
+      industrial: 40,
+    },
+  };
+}
+
+export function calculateAverageCoverage(coverage: number[][]): number {
+  let total = 0;
+  let count = 0;
+  for (const row of coverage) {
+    for (const value of row) {
+      total += value;
+      count++;
+    }
+  }
+  return count > 0 ? total / count : 0;
+}
+
+// PERF: Update budget costs based on buildings - single pass through grid
+export function updateBudgetCosts(grid: Tile[][], budget: Budget): Budget {
+  const newBudget = { ...budget };
+  
+  let policeCount = 0;
+  let fireCount = 0;
+  let hospitalCount = 0;
+  let schoolCount = 0;
+  let universityCount = 0;
+  let parkCount = 0;
+  let powerCount = 0;
+  let waterCount = 0;
+  let roadCount = 0;
+  let subwayTileCount = 0;
+  let subwayStationCount = 0;
+
+  // PERF: Single pass through grid instead of two separate loops
+  for (const row of grid) {
+    for (const tile of row) {
+      // Count subway tiles
+      if (tile.hasSubway) subwayTileCount++;
+      
+      // Count building types using switch for jump table optimization
+      switch (tile.building.type) {
+        case 'police_station': policeCount++; break;
+        case 'fire_station': fireCount++; break;
+        case 'hospital': hospitalCount++; break;
+        case 'school': schoolCount++; break;
+        case 'university': universityCount++; break;
+        case 'park': parkCount++; break;
+        case 'park_large': parkCount++; break;
+        case 'tennis': parkCount++; break;
+        case 'power_plant': powerCount++; break;
+        case 'water_tower': waterCount++; break;
+        case 'road': roadCount++; break;
+        case 'subway_station': subwayStationCount++; break;
+      }
+    }
+  }
+
+  newBudget.police.cost = policeCount * 50;
+  newBudget.fire.cost = fireCount * 50;
+  newBudget.health.cost = hospitalCount * 100;
+  newBudget.education.cost = schoolCount * 30 + universityCount * 100;
+  newBudget.transportation.cost = roadCount * 2 + subwayTileCount * 3 + subwayStationCount * 25;
+  newBudget.parks.cost = parkCount * 10;
+  newBudget.power.cost = powerCount * 150;
+  newBudget.water.cost = waterCount * 75;
+
+  return newBudget;
+}
+
+// Calculate city stats
+// effectiveTaxRate is the lagged tax rate used for demand calculations
+export function calculateStats(grid: Tile[][], size: number, budget: Budget, taxRate: number, effectiveTaxRate: number, services: ServiceCoverage): Stats {
+  let population = 0;
+  let jobs = 0;
+  let totalPollution = 0;
+  let residentialZones = 0;
+  let commercialZones = 0;
+  let industrialZones = 0;
+  let developedResidential = 0;
+  let developedCommercial = 0;
+  let developedIndustrial = 0;
+  let totalLandValue = 0;
+  let treeCount = 0;
+  let waterCount = 0;
+  let parkCount = 0;
+  let subwayTiles = 0;
+  let subwayStations = 0;
+  let railTiles = 0;
+  let railStations = 0;
+  
+  // Special buildings that affect demand
+  let hasAirport = false;
+  let hasCityHall = false;
+  let hasSpaceProgram = false;
+  let stadiumCount = 0;
+  let museumCount = 0;
+  let hasAmusementPark = false;
+
+  // Count everything
+  for (let y = 0; y < size; y++) {
+    for (let x = 0; x < size; x++) {
+      const tile = grid[y][x];
+      const building = tile.building;
+
+      // Apply subway commercial boost to jobs (tiles with subway get 15% boost to commercial jobs)
+      let jobsFromTile = building.jobs;
+      if (tile.hasSubway && tile.zone === 'commercial') {
+        jobsFromTile = Math.floor(jobsFromTile * 1.15);
+      }
+      
+      population += building.population;
+      jobs += jobsFromTile;
+      totalPollution += tile.pollution;
+      totalLandValue += tile.landValue;
+
+      if (tile.zone === 'residential') {
+        residentialZones++;
+        if (building.type !== 'grass' && building.type !== 'empty') developedResidential++;
+      } else if (tile.zone === 'commercial') {
+        commercialZones++;
+        if (building.type !== 'grass' && building.type !== 'empty') developedCommercial++;
+      } else if (tile.zone === 'industrial') {
+        industrialZones++;
+        if (building.type !== 'grass' && building.type !== 'empty') developedIndustrial++;
+      }
+
+      if (building.type === 'tree') treeCount++;
+      if (building.type === 'water') waterCount++;
+      if (building.type === 'park' || building.type === 'park_large') parkCount++;
+      if (building.type === 'tennis') parkCount++; // Tennis courts count as parks
+      if (tile.hasSubway) subwayTiles++;
+      if (building.type === 'subway_station') subwayStations++;
+      if (building.type === 'rail' || tile.hasRailOverlay) railTiles++;
+      if (building.type === 'rail_station') railStations++;
+      
+      // Track special buildings (only count if construction is complete)
+      if (building.constructionProgress === undefined || building.constructionProgress >= 100) {
+        if (building.type === 'airport') hasAirport = true;
+        if (building.type === 'city_hall') hasCityHall = true;
+        if (building.type === 'space_program') hasSpaceProgram = true;
+        if (building.type === 'stadium') stadiumCount++;
+        if (building.type === 'museum') museumCount++;
+        if (building.type === 'amusement_park') hasAmusementPark = true;
+      }
+    }
+  }
+
+  // Calculate demand - subway network boosts commercial demand
+  // Tax rate affects demand as BOTH a multiplier and additive modifier:
+  // - Multiplier: At 100% tax, demand is reduced to 0 regardless of other factors
+  // - Additive: Small bonus/penalty around the base rate for fine-tuning
+  // Base tax rate is 9%, so we calculate relative to that
+  // Uses effectiveTaxRate (lagged) so changes don't impact demand immediately
+  
+  // Tax multiplier: 1.0 at 0% tax, ~1.0 at 9% tax, 0.0 at 100% tax
+  // This ensures high taxes dramatically reduce demand regardless of other factors
+  const taxMultiplier = Math.max(0, 1 - (effectiveTaxRate - 9) / 91);
+  
+  // Small additive modifier for fine-tuning around base rate
+  // At 9% tax: 0. At 0% tax: +18. At 20% tax: -22
+  const taxAdditiveModifier = (9 - effectiveTaxRate) * 2;
+  
+  const subwayBonus = Math.min(20, subwayTiles * 0.5 + subwayStations * 3);
+  
+  // Rail network bonuses - affects commercial (passenger rail, accessibility) and industrial (freight transport)
+  // Rail stations have bigger impact than raw track count since they represent actual service
+  // Industrial gets a stronger bonus as freight rail is critical for factories/warehouses
+  const railCommercialBonus = Math.min(12, railTiles * 0.15 + railStations * 4);
+  const railIndustrialBonus = Math.min(18, railTiles * 0.25 + railStations * 6);
+  
+  // Special building bonuses
+  // Airport: Major boost to commercial (business travel) and industrial (cargo/logistics)
+  const airportCommercialBonus = hasAirport ? 15 : 0;
+  const airportIndustrialBonus = hasAirport ? 10 : 0;
+  
+  // City Hall: Modest boost to all demand (legitimacy, attracts businesses and residents)
+  const cityHallResidentialBonus = hasCityHall ? 8 : 0;
+  const cityHallCommercialBonus = hasCityHall ? 10 : 0;
+  const cityHallIndustrialBonus = hasCityHall ? 5 : 0;
+  
+  // Space Program: Big boost to industrial (high-tech sector), modest boost to residential (prestige)
+  const spaceProgramResidentialBonus = hasSpaceProgram ? 10 : 0;
+  const spaceProgramIndustrialBonus = hasSpaceProgram ? 20 : 0;
+  
+  // Stadium: Boost to commercial (entertainment, visitors, sports bars)
+  const stadiumCommercialBonus = Math.min(20, stadiumCount * 12);
+  
+  // Museum: Boost to commercial (tourism) and residential (culture/quality of life)
+  const museumCommercialBonus = Math.min(15, museumCount * 8);
+  const museumResidentialBonus = Math.min(10, museumCount * 5);
+  
+  // Amusement Park: Big boost to commercial (tourism, entertainment)
+  const amusementParkCommercialBonus = hasAmusementPark ? 18 : 0;
+  
+  // Calculate base demands from economic factors
+  const baseResidentialDemand = (jobs - population * 0.7) / 18;
+  const baseCommercialDemand = (population * 0.3 - jobs * 0.3) / 4 + subwayBonus;
+  const baseIndustrialDemand = (population * 0.35 - jobs * 0.3) / 2.0;
+  
+  // Add special building bonuses to base demands
+  const residentialWithBonuses = baseResidentialDemand + cityHallResidentialBonus + spaceProgramResidentialBonus + museumResidentialBonus;
+  const commercialWithBonuses = baseCommercialDemand + airportCommercialBonus + cityHallCommercialBonus + stadiumCommercialBonus + museumCommercialBonus + amusementParkCommercialBonus + railCommercialBonus;
+  const industrialWithBonuses = baseIndustrialDemand + airportIndustrialBonus + cityHallIndustrialBonus + spaceProgramIndustrialBonus + railIndustrialBonus;
+  
+  // Apply tax effect: multiply by tax factor, then add small modifier
+  // The multiplier ensures high taxes crush demand; the additive fine-tunes at normal rates
+  const residentialDemand = Math.min(100, Math.max(-100, residentialWithBonuses * taxMultiplier + taxAdditiveModifier));
+  const commercialDemand = Math.min(100, Math.max(-100, commercialWithBonuses * taxMultiplier + taxAdditiveModifier * 0.8));
+  const industrialDemand = Math.min(100, Math.max(-100, industrialWithBonuses * taxMultiplier + taxAdditiveModifier * 0.5));
+
+  // Calculate income and expenses
+  const income = Math.floor(population * taxRate * 0.1 + jobs * taxRate * 0.05);
+  
+  let expenses = 0;
+  expenses += Math.floor(budget.police.cost * budget.police.funding / 100);
+  expenses += Math.floor(budget.fire.cost * budget.fire.funding / 100);
+  expenses += Math.floor(budget.health.cost * budget.health.funding / 100);
+  expenses += Math.floor(budget.education.cost * budget.education.funding / 100);
+  expenses += Math.floor(budget.transportation.cost * budget.transportation.funding / 100);
+  expenses += Math.floor(budget.parks.cost * budget.parks.funding / 100);
+  expenses += Math.floor(budget.power.cost * budget.power.funding / 100);
+  expenses += Math.floor(budget.water.cost * budget.water.funding / 100);
+
+  // Calculate ratings
+  const avgPoliceCoverage = calculateAverageCoverage(services.police);
+  const avgFireCoverage = calculateAverageCoverage(services.fire);
+  const avgHealthCoverage = calculateAverageCoverage(services.health);
+  const avgEducationCoverage = calculateAverageCoverage(services.education);
+
+  const safety = Math.min(100, avgPoliceCoverage * 0.7 + avgFireCoverage * 0.3);
+  const health = Math.min(100, avgHealthCoverage * 0.8 + (100 - totalPollution / (size * size)) * 0.2);
+  const education = Math.min(100, avgEducationCoverage);
+  
+  const greenRatio = (treeCount + waterCount + parkCount) / (size * size);
+  const pollutionRatio = totalPollution / (size * size * 100);
+  const environment = Math.min(100, Math.max(0, greenRatio * 200 - pollutionRatio * 100 + 50));
+
+  const jobSatisfaction = jobs >= population ? 100 : (jobs / (population || 1)) * 100;
+  const happiness = Math.min(100, (
+    safety * 0.15 +
+    health * 0.2 +
+    education * 0.15 +
+    environment * 0.15 +
+    jobSatisfaction * 0.2 +
+    (100 - taxRate * 3) * 0.15
+  ));
+
+  return {
+    population,
+    jobs,
+    money: 0, // Will be updated from previous state
+    income,
+    expenses,
+    happiness,
+    health,
+    education,
+    safety,
+    environment,
+    demand: {
+      residential: residentialDemand,
+      commercial: commercialDemand,
+      industrial: industrialDemand,
+    },
+  };
+}
